TFM_CTF_FINAL/requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
protobuf==4.21.12
Pygments==2.16.1
six==1.16.0
Werkzeug==2.3.7

pip~=21.3.1
wheel~=0.37.1
setuptools~=60.2.0
zipp~=3.16.2
bleach~=6.0.0